rdp: Move audio processing to front-end
Move the bulk of the audio related code into the compositor front end, and
instead put callbacks to setup functions into the rdp backend config
structure.

There are some changes to logging because we no longer have access to the
backend's logging scopes. Instead we create rdp-audio and rdp-audio-in
scopes.

Signed-off-by: Derek Foreman <derek.foreman@collabora.com>