rdp: Simplify clipboard_client_format_list
isPublished is always set if source is valid and never set if it isn't,
so we can just throw away isPublished and test source directly. This
also shows that the clipboard_data_source_unref block can never evaluate
to true, so we can remove that too.

For future reference, that was the only call site for
clipboard_data_source_unref that wasn't from the compositor main thread.

Signed-off-by: Derek Foreman <derek.foreman@collabora.com>