rdp: Put wl_array read from fd into a function
This is a pretty unusual use of wl_array, so let's compartmentalize it as
much as possible into a function.

There's a slight functional change here - we always ensure that there's
enough space for a null terminator after the read, but that slack byte
will be uninitialized and the consumers of the data may need to set it
to a null.

We also add EINTR handling to the read.

Signed-off-by: Derek Foreman <derek.foreman@collabora.com>